test(core): refactor tests for TextFieldGroup
